requirements.txt
pandas
biopython
biotite
requests